updated cancel button to match figma
